requirements.txt
streamlit 
Pillow
pytesseract
spacy
scikit-learn
pandas
numpy
python-magic